Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -21,7 +21,7 @@ int main() {
         printf("%d",mul);
         break;
         case '/': if (b != 0) {  // Prevent division by zero
-                div = (float)a / b;  // Cast to float for correct division
+                div = a / b;  // Cast to float for correct division
                 printf("%f", div);
             } else {
                 printf("error");
